Refactoring code into different modules
