+++ heart.py
@@ -1,0 +1,8 @@
+#
+# Heart sprite
+#
+import arcade
+
+class Heart(arcade.Sprite):
+    def __init__(self, filename, scaling):
+        super().__init__(filename, scaling)
